+++ finalProject/server/routes/users.js
@@ -1,6 +1,10 @@
 var express = require('express');
 var router = express.Router();
 const User = require('../models/user');
+const Collection = require('../models/collection');
+const bcrypt = require('bcryptjs');
+const { default: mongoose } = require('mongoose');
+const Forum = require('../models/forum');
 
 // GET all users, paginated
 router.get('/users', async (req, res) => {
@@ -28,8 +32,9 @@ router.get('/users', async (req, res) => {
 // Create a new user
 router.post('/register', async (req, res) => {
   try {
-    console.log(req.body);
-    const user = await User.create(req.body);
+    const { password } = req.body;
+    const newPassword = await bcrypt.hash(password, 10);
+    const user = await User.create({ ...req.body, password: newPassword });
     res.status(201).json(user);
   } catch (error) {
     console.error("User creation failed: ", error);
@@ -39,6 +44,9 @@ router.post('/register', async (req, res) => {
 
 router.get('/users/:id', async (req, res) => {
   try {
+    const userId = new mongoose.Types.ObjectId(req.params.id);
+    console.log(userId);
+
     const user = await User.findById(req.params.id);
     if (!user) {
       return res.status(404).json({ error: 'User not found' });
@@ -52,6 +60,12 @@ router.get('/users/:id', async (req, res) => {
 // Update a user
 router.put('/users/:id', async (req, res) => {
   try {
+    const userId = new mongoose.Types.ObjectId(req.params.id);
+
+    if(userId.toString() !== req.user._id.toString()) {
+      return res.status(401).json({ error: 'Unauthorized' });
+    }
+
     const user = await User.findByIdAndUpdate(req.params.id, req.body, { new: true });
     if (!user) {
       return res.status(404).json({ error: 'User not found' });
@@ -65,10 +79,27 @@ router.put('/users/:id', async (req, res) => {
 // Delete a user
 router.delete('/users/:id', async (req, res) => {
   try {
+    const userId = new mongoose.Types.ObjectId(req.params.id);
+
+    console.log(req.user.role);
+    if((userId.toString() !== req.user._id.toString()) && req.user.role !== 'admin') {
+      return res.status(401).json({ error: 'Unauthorized' });
+    }
+
+    const forums = await Forum.deleteMany({ userId: req.params.id });
+
+    const collection = await Collection.findOneAndDelete({ userId: req.params.id });
+
+    if (!collection) {
+      return res.status(404).json({ error: 'Collection not found' });
+    }
+
     const user = await User.findByIdAndDelete(req.params.id);
+
     if (!user) {
       return res.status(404).json({ error: 'User not found' });
     }
+    
     res.json(user);
   } catch (error) {
     res.status(500).json({ error: error.message });
